+++ resources/lang/de/validation.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 return [
     'email' => [
-        'domain_not_allowed'  => 'Die Domain :domain ist nicht zulässig. Bitte versuchen Sie es mit einer anderen E-Mail-Adresse.',
-        'invalid'             => 'Dieses :attribute muss eine gültige E-Mail-Adresse sein.',
-        'risky'               => 'Dieses :attribute kann möglicherweise nicht sicher zugestellt werden. Bitte geben Sie ein anderes an.',
-        'service_unavailable' => 'Wir können dieses :attribute im Moment nicht prüfen. Bitte versuchen Sie es später erneut.',
-        'undeliverable'       => 'Dieses :attribute ist ungültig und kann nicht zugestellt werden. Bitte geben Sie ein gültiges an.',
+        'domain_not_allowed' => 'Die Domain :domain ist nicht zulässig. Bitte versuchen Sie es mit einer anderen E-Mail-Adresse.',
+        'invalid'            => 'Dieses :attribute muss eine gültige E-Mail-Adresse sein.',
+        'risky'              => 'Dieses :attribute kann möglicherweise nicht sicher zugestellt werden. Bitte geben Sie ein anderes an.',
+        'undeliverable'      => 'Dieses :attribute ist ungültig und kann nicht zugestellt werden. Bitte geben Sie ein gültiges an.',
+        'unverifiable'       => 'Wir können dieses :attribute im Moment nicht prüfen. Bitte versuchen Sie es später erneut.',
     ],
 ];