+++ resources/lang/en/validation.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 return [
     'email' => [
-        'domain_not_allowed'  => 'The :domain domain is not allowed. Please try another email address.',
-        'invalid'             => 'The :attribute must be a valid email address.',
-        'risky'               => 'This :attribute may not be safely deliverable. Please provide another one.',
-        'service_unavailable' => 'We are unable to check this :attribute right now. Please try again later.',
-        'undeliverable'       => 'This :attribute is invalid and cannot be delivered. Please provide a valid one.',
+        'domain_not_allowed' => 'The :domain domain is not allowed. Please try another email address.',
+        'invalid'            => 'The :attribute must be a valid email address.',
+        'risky'              => 'This :attribute may not be safely deliverable. Please provide another one.',
+        'undeliverable'      => 'This :attribute is invalid and cannot be delivered. Please provide a valid one.',
+        'unverifiable'       => 'We are unable to check this :attribute right now. Please try again later.',
     ],
 ];